add verify function on Trezor
