+++ app/stores/ada/TrezorSendStore.js
@@ -1,14 +1,10 @@
 // @flow
 import { action, observable } from 'mobx';
-import { defineMessages } from 'react-intl';
 import TrezorConnect from 'trezor-connect';
 
 import Store from '../base/Store';
 import environment from '../../environment';
 import LocalizedRequest from '../lib/LocalizedRequest';
-
-import LocalizableError, { UnexpectedError } from '../../i18n/LocalizableError';
-import globalMessages from '../../i18n/global-messages';
 
 import type {
   CreateTrezorSignTxDataRequest,
@@ -21,13 +17,10 @@ import {
   Logger,
   stringifyError,
 } from '../../utils/logging';
-
-const messages = defineMessages({
-  signTxError101: {
-    id: 'wallet.send.trezor.error.101',
-    defaultMessage: '!!!Signing cancelled on Trezor device. Please retry.',
-  },
-});
+import {
+  convertToLocalizableError
+} from '../../domain/TrezorLocalizedError';
+import LocalizableError from '../../i18n/LocalizableError';
 
 /** Note: Handles Trezor Signing */
 export default class TrezorSendStore extends Store {
@@ -99,7 +92,7 @@ export default class TrezorSendStore extends Store {
 
     } catch (error) {
       Logger.error('TrezorSendStore::_sendUsingTrezor error: ' + stringifyError(error));
-      this._setError(this._convertToLocalizableError(error));
+      this._setError(convertToLocalizableError(error));
     } finally {
       this.createTrezorSignTxDataRequest.reset();
       this.broadcastTrezorSignedTxRequest.reset();
@@ -129,47 +122,6 @@ export default class TrezorSendStore extends Store {
     }
 
     Logger.info('SUCCESS: ADA sent using Trezor SignTx');
-  }
-
-  /** Converts error(from API or Trezor API) to LocalizableError */
-  _convertToLocalizableError = (error: any): LocalizableError => {
-    let localizableError: ?LocalizableError = null;
-
-    if (error instanceof LocalizableError) {
-      // It means some API Error has been thrown
-      localizableError = error;
-    } else if (error && error.message) {
-      // Trezor device related error happend, convert then to LocalizableError
-      switch (error.message) {
-        case 'Iframe timeout':
-          localizableError = new LocalizableError(globalMessages.trezorError101);
-          break;
-        case 'Permissions not granted':
-          localizableError = new LocalizableError(globalMessages.hwError101);
-          break;
-        case 'Cancelled':
-        case 'Popup closed':
-          localizableError = new LocalizableError(globalMessages.trezorError103);
-          break;
-        case 'Signing cancelled':
-          localizableError = new LocalizableError(messages.signTxError101);
-          break;
-        default:
-          /** we are not able to figure out why Error is thrown
-            * make it, Something unexpected happened */
-          Logger.error(`TrezorSendStore::_convertToLocalizableError::error: ${error.message}`);
-          localizableError = new UnexpectedError();
-          break;
-      }
-    }
-
-    if (!localizableError) {
-      /** we are not able to figure out why Error is thrown
-        * make it, Something unexpected happened */
-      localizableError = new UnexpectedError();
-    }
-
-    return localizableError;
   }
 
   _cancel = (): void => {
